fix: refactor credhub_setup
Moved various things to sub-packages for better testability.
