+++ deploy/containers/credhub_setup/cc/doc.go
@@ -1,0 +1,2 @@
+// Package cc contains the code dealing with the CF cloud controller API.
+package cc
