+++ deploy/containers/credhub_setup/quarks/doc.go
@@ -1,0 +1,2 @@
+// Package quarks contains methods dealing with the quarks environment
+package quarks
