+++ deploy/containers/credhub_setup/uaa/auth_test.go
@@ -1,4 +1,4 @@
-package main
+package uaa
 
 import (
 	"bytes"
@@ -17,30 +17,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-func TestMakeHTTPClientWithCA(t *testing.T) {
-	t.Parallel()
-	server := httptest.NewTLSServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		fmt.Fprintf(w, "ok")
-	}))
-	defer server.Close()
-
-	serverURL, err := url.Parse(server.URL)
-	require.NoError(t, err, "error parsing server url")
-
-	certBytes := bytes.Buffer{}
-	pem.Encode(&certBytes, &pem.Block{
-		Type:  "CERTIFICATE",
-		Bytes: server.Certificate().Raw,
-	})
-	client, err := makeHTTPClientWithCA(serverURL.Hostname(), certBytes.Bytes())
-	require.NoError(t, err, "failed to make HTTP client")
-
-	resp, err := client.Get(server.URL)
-	require.NoError(t, err, "error fetching from test server")
-	require.GreaterOrEqual(t, resp.StatusCode, 200, "unexpected status: %s", resp.Status)
-	require.Less(t, resp.StatusCode, 300, "unexpected status: %s", resp.Status)
-}
 
 type mockAuthServer struct {
 	*http.ServeMux
