+++ deploy/containers/credhub_setup/uaa/doc.go
@@ -1,0 +1,3 @@
+// package uaa contains functions dealing with UAA authentication.
+
+package uaa
